Allow partial class interception of client builder methods
This will make implementing emulator support simpler, for example.

An alternative possibly design would be to have a partial method
"returning" (via a ref parameter) another builder of the same type.
We could then, then we could call Build or BuildAsync on that
builder (if not null). This would mean only implementing a single
partial method in partial classes - but would prohibit any
customization requiring asynchrony.

The scheme in this commit is still not massively onerous for partial
class implementers, but gives more freedom for future expansion.
Note that all of this can change in the generator later, in that
it's a breaking change for generator consumers, but not for end user
developers, as the public API surface remains the same
(Build/BuildAsync).