fix: Use gapic_metadata.json to determine where to copy extra files
This uses the fact that the metadata file is generated from the
first service to be specified to the GAPIC generator, which is
expected to be "the one for the client library".

Other services may be generated (e.g. for mixins) which currently
ends up with files in the wrong place.